requirements.txt
blessings==1.6
bpython==0.15
curtsies==0.2.6
Django==1.9.7
djangorestframework==3.3.3
greenlet==0.4.9
psycopg2==2.6.1
Pygments==2.1.3
requests==2.10.0
six==1.10.0
wcwidth==0.1.6
